ロジックのカスタムフックに切り出し
