+++ components/list/Dialog.tsx
@@ -9,7 +9,6 @@ import {
 import { Button } from "@/components/ui/button";
 
 import { Toaster } from "@/components/ui/sonner";
-import { toast } from "sonner";
 import { Input } from "@/components/ui/input";
 import { Textarea } from "@/components/ui/textarea";
 import {
@@ -22,6 +21,7 @@ import {
 import { Task } from "@/types/tasks";
 import StatusIcon from "../StatusIcon";
 import EmergencyIcon from "../EmergencyIcon";
+import { useHandleDialog } from "./hooks/useHandleDialog";
 
 type DialogProps = {
   selectedTask: Task | null;
@@ -38,82 +38,14 @@ const Dialog: React.FC<DialogProps> = ({
   setIsDialogOpen,
   setTasks,
 }) => {
-  //モーダルを開く処理=======================================================================
-  const handleChange = (field: keyof Task, value: string) => {
-    if (selectedTask) {
-      setSelectedTask({ ...selectedTask, [field]: value });
-    }
-  };
-
-  //更新用処理=======================================================================
-  const handleSave = async () => {
-    if (selectedTask) {
-      try {
-        const response = await fetch(
-          `/api/task/put?selectedTaskId=${selectedTask.id}`,
-          {
-            method: "PUT",
-            headers: {
-              "Content-Type": "application/json",
-            },
-            body: JSON.stringify(selectedTask),
-          }
-        );
-
-        if (!response.ok) {
-          throw new Error("Failed to update task");
-        }
-
-        // 成功レスポンスから更新済みタスクを取得
-        const { data: updatedTask } = await response.json();
-
-        // ローカルタスクリストを更新
-        setTasks((prevTasks) =>
-          prevTasks.map((task) =>
-            task.id === updatedTask.id ? updatedTask : task
-          )
-        );
-
-        setIsDialogOpen(false);
-        toast.success("Task updated successfully!");
-      } catch (error) {
-        console.error(error);
-        toast.error("Error updating task");
-      }
-    }
-  };
   //==========================================================================
 
-  //削除用処理=======================================================================
-
-  const handleDelete = async () => {
-    if (selectedTask) {
-      try {
-        const response = await fetch(
-          `/api/task/delete?selectedTaskId=${selectedTask.id}`,
-          {
-            method: "DELETE",
-          }
-        );
-
-        if (!response.ok) {
-          throw new Error("Failed to delete task");
-        }
-
-        setTasks((prevTasks) =>
-          prevTasks.filter((task) => task.id !== selectedTask.id)
-        );
-
-        setIsDialogOpen(false);
-        toast.success("Task deleted successfully!");
-      } catch (error) {
-        console.error(error);
-        toast.error("Error deleting task");
-      }
-    }
-  };
-
-  //==========================================================================
+  const { handleChange, handleSave, handleDelete } = useHandleDialog({
+    selectedTask,
+    setSelectedTask,
+    setIsDialogOpen,
+    setTasks,
+  });
 
   return (
     <>
