+++ components/list/List.tsx
@@ -1,4 +1,4 @@
-import React, { Dispatch, SetStateAction, useEffect, useState } from "react";
+import React, { Dispatch, SetStateAction } from "react";
 
 import {
   Table,
@@ -14,6 +14,8 @@ import { Task } from "@/types/tasks";
 import { Category } from "@/types/category";
 import Dialog from "./Dialog";
 import { Button } from "../ui/button";
+import { useGetTask } from "./hooks/useGetTask";
+import { useTaskSort } from "./hooks/useTaskSort";
 export const List = ({
   isAddTask,
   selectCategory,
@@ -25,63 +27,25 @@ export const List = ({
   tasks: Task[];
   setTasks: Dispatch<SetStateAction<Task[]>>;
 }) => {
-  const [selectedTask, setSelectedTask] = useState<Task | null>(null); // 選択したtaskを保持
-  const [isDialogOpen, setIsDialogOpen] = useState<boolean>(false); // ダイアログフラグ
-  const [loading, setLoading] = useState<boolean>(true); // ローディング状態
-  // APIからデータを取得==========================================================-
-  useEffect(() => {
-    async function fetchTasks() {
-      setLoading(true); // データ取得開始
-      try {
-        // selectCategoryのIDをクエリパラメータとして追加
-        const url = `/api/task/get?categoryId=${selectCategory.id}`;
-
-        const response = await fetch(url);
-        if (!response.ok) {
-          const errorData = await response.json();
-          console.error("API Error:", errorData);
-          throw new Error(errorData.details || "Failed to fetch data");
-        }
-        const data: { tasks: Task[] } = await response.json();
-        setTasks(data.tasks);
-      } catch (error) {
-        console.error("データの取得中にエラーが発生しました:", error);
-      } finally {
-        setLoading(false); // データ取得完了
-      }
-    }
-
-    fetchTasks();
-  }, [isAddTask, selectCategory]); //タスク作成フラグまたは、選択しているカテゴリの値が変わった時にデータを再取得する
   //==================================================================================
 
+  const {
+    selectedTask,
+    setSelectedTask,
+    isDialogOpen,
+    setIsDialogOpen,
+    loading,
+  } = useGetTask({ selectCategory, setTasks, isAddTask });
+
+  const { handleHeaderClick } = useTaskSort({
+    tasks,
+    setTasks,
+  });
+
   // ダブルクリックでタスクを選択し、ダイアログを開く
-  const handleRowDoubleClick = (task: Task) => {
+  const handleTaskClick = (task: Task) => {
     setSelectedTask(task); // 選択されたタスクを設定
     setIsDialogOpen(true); // ダイアログを開く
-  };
-
-  //sort機能==========================================================================
-  const [sortColumn, setSortColumn] = useState<keyof Task | null>(null); // ソート対象の列
-  const [sortOrder, setSortOrder] = useState<"asc" | "desc">("asc"); // ソート順
-
-  const handleHeaderClick = (column: keyof Task) => {
-    if (sortColumn === column) {
-      // 同じ列を再度クリックした場合は並び順を切り替え
-      setSortOrder((prevOrder) => (prevOrder === "asc" ? "desc" : "asc"));
-    } else {
-      // 別の列をクリックした場合は新しい列で昇順ソート
-      setSortColumn(column);
-      setSortOrder("asc");
-    }
-
-    // 並び替え処理
-    const sortedTasks = [...tasks].sort((a, b) => {
-      if (a[column] < b[column]) return sortOrder === "asc" ? -1 : 1;
-      if (a[column] > b[column]) return sortOrder === "asc" ? 1 : -1;
-      return 0;
-    });
-    setTasks(sortedTasks);
   };
 
   //=================================================================-
@@ -112,10 +76,7 @@ export const List = ({
             </TableRow>
           ) : tasks.length > 0 ? (
             tasks.map((item, index) => (
-              <TableRow
-                key={index}
-                onDoubleClick={() => handleRowDoubleClick(item)}
-              >
+              <TableRow key={index} onClick={() => handleTaskClick(item)}>
                 <TableCell>{item.title}</TableCell>
                 <TableCell>
                   <EmergencyIcon emergency={item.emergency} />
